refactor(frontend): tooltip moved to responding folder
